requirements.txt
Pillow~=9.1.0
pandas~=1.3.5
requests~=2.27.1
python-dotenv~=0.20.0
fastapi~=0.75.2
uvicorn[standard]~=0.17.6
pymongo~=4.0.2
torch~=1.11.0
torchvision~=0.12.0
minio~=7.1.7
tqdm~=4.64.0
fic>=0.8
opencv-python-headless~=4.5.5.64
python-multipart~=0.0.5
fastapi_jwt_auth~=0.5.0
pydantic~=1.8.2
